Fix the compiler warnings about 64->32 bits.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@297935 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native2/common/jk_worker_status.c
@@ -114,9 +114,9 @@ static void jk2_worker_status_displayStat(jk_env_t *env, jk_ws_service_t *s,
         s->jkprintf(env, s, "<td>%ld</td>\n",
                     (long)(stat->endTime - stat->startTime) );
         
-        totalTime += stat->totalTime;
+        totalTime += (long)stat->totalTime;
         if( maxTime < stat->maxTime )
-            maxTime=stat->maxTime;
+            maxTime = (long)stat->maxTime;
     }
 #endif
     s->jkprintf(env, s, "</tr>\n");
